move pkgs twitter, slack to root dir
